backend/requirements.txt
# Django y REST framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Base de datos
psycopg2-binary==2.9.9

# Cache y Redis
django-redis==5.4.0
redis==5.0.1

# Autenticación
djangorestframework-simplejwt==5.3.1

# Plaid (para integración bancaria)
plaid-python==18.1.0

# Utilidades
python-dotenv==1.0.0
requests==2.31.0

# Desarrollo y testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov>=4.1.0,<4.2.0

